Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/2994071

+++ src/trace_processor/containers/bit_vector.cc
@@ -395,7 +395,8 @@ void BitVector::SelectBits(const BitVector& mask_bv) {
 
   // Loop post-condition: we must have written as many words as is required
   // to store |set_bits_in_mask|.
-  PERFETTO_DCHECK(out_word - words_.data() <= WordCount(set_bits_in_mask));
+  PERFETTO_DCHECK(static_cast<uint32_t>(out_word - words_.data()) <=
+                  WordCount(set_bits_in_mask));
 
   // Resize the BitVector to equal to the number of elements in the  mask we
   // calculated at the start of the loop.
